[NEW] = LB_ELECTROHYDRODYNAMICS: charge coupling scheme for Lattice-Boltzmann simulations. See publication Hickey et. al. (2010)
Added new feature:
LB_ELECTROHYDRODYNAMICS

With this, you can use a coupling scheme to simulate electrostatics and hydrodynamics
for electrophoresis without long-range interactions, simply in Lattice-Boltzmann.

For each particle, a new 3D-parameter mu_E can be set and is stored in Variable p->p.mu_E[3]

This parameter is added to the LB-force routine.

Documentation yet to come.