This is a small guide to the new naming convention in ESPResSo.
Tcl commands parsing functions in ESPResSo were somehow heterogeneous
both in logical structure and in naming convention: I tried to make
some order as explained below. Summarizing, now every function
involving the Tcl interpreter(*) has one of the following forms

                 tclcommand_SOMETHING
                 tclcallback_SOMETHING
                 tclprint_SOMETHING

Please notice that this is still a test version: the naming convention
can still (will) be changed, and not every combination of #defines
in myconfig.h has been tried, so some problems might still pop out.
Feel free to contact me (sega@icp.uni-stuttgart.de) (a) about
problematic as well as successful merges -good news are always
welcome- with your actual branch (b) if some code got broken (c)
to suggest improvements, or changes in the naming scheme.

(*) besides very few exceptions like: functions commented out, not
used anywhere in the code, or marked as experimental (hence, which
are supposed anyway to be cleaned/checked),  the
mpi_gather_runtime_errors(), on_program_start() appinit(),
parse_*_list(), and init_tcl() functions. Some UWerr functions using
the interpreter were also not changed, because new tcl commands are
created within the module, and need probably a different treatment.

The details of the naming convention follows.

*********************************************************************
		  TCL COMMANDS AND RELATED
*********************************************************************
All C functions associated to an ESPResSo tcl command have now the
form

            tclcommand_COMMAND()

where COMMAND corresponds without any exception to the name of the
corresponding tcl command.  This should make rather easy to spot
them in the code.  For example,

	    grep "tclcommand_part\>" *.[ch]

searches for the references to the "part" command.

----------------------------------------------------------------------

When the first argument of a command needs to be parsed, as in the
case of 'inter coulomb', the name of the parsing function is always

           tclcommand_COMMAND_parse_SUBCOMMAND()

again, with strict correspondence to the tcl interface (this was
not always the case before).  Suppose you want to search for all
possible ways of calling the 'inter' command, then just

	   grep "tclcommand_inter_parse_[a-z]\+" *.[ch]

----------------------------------------------------------------------

The functions parsing one level deeper are called

           tclcommand_COMMAND_SUBCOMMAND_parse_ARGUMENT()

so, for example, the function which parses 'inter coulomb p3m' is

	   tclcommand_inter_coulomb_parse_p3m()

After this level, the way parsing functions are organized is too
heterogeneous to find a unifiying naming convention without rewriting
code, hence, all subsequent options are just appended after an
underscore:

	   tclcommand_COMMAND_SUBCOMMAND_parse_ARGUMENT_OPTION()

Thus, the function that parses 'inter coulomb p3m tune' is

	   tclcommand_inter_coulomb_parse_p3m_tune()

The strict correspondence to the tcl command is also lost at this
level, as for example the tclcommand_inter_coulomb_parse_p3m_tune()
function is used to parse both the 'tune' and the 'tunev2' options.

----------------------------------------------------------------------
In ESPResSo, some functions involving the tcl interpreter are used
just to print out some information, and do not parse anything (the
opposite often happens). Even though in many cases they are also
performing some operation but jut I/O, they have been called

           tclcommand_COMMAND_print_SOMETHING

like in

           tclcommand_inter_print_all
or
           tclcommand_inter_print_bonded

Not in all cases it was possible to relate SOMETHING to a tcl
command option.

**********************************************************************
			   CALLBACKS
**********************************************************************

All callbacks have now either the form

           tclcallback_NAME

where NAME is the tcl name of the variable addressed, or they point
to some default read-only message, like  tclcallback_ro() or
tclcallback_thermo_ro().

**********************************************************************
		printXYZToResult() FUNCTIONS
**********************************************************************

All functions printXYZToResult() have been renamed to
tclprint_to_result_XYZ()

**********************************************************************
			    MISC
**********************************************************************

In addition, going through the code, all functions (not just parsers)
which are not called anywhere have been marked with a comment like

/* TODO: this function is not used anywhere. To be removed? */

----------------------------------------------------------------------

This patch includes also some minor changes to clean the code.

**********************************************************************
EOF