Update to file samples/correlation.tcl and to UG
The sample file was badly outdated, relating to an ancient
implementation of the correlator. Still parts of it need to
be re-done because I do not get what they were supposed to do.

UG section on correlator and analysis was updated and made more
complete. Still not finished, though.

Also commented out some forgotten debug messages printed to stderr
by the correlator.