Docs of more observables, removed some confusing ones
Removed very problem-specific and confusing observable nearest_neighbour_conditional.

Commented out the structure_factor observable. The current implementation has not
been finished and produces a segfault. It should be kept in the code and the implementation
finished later.

The corresponding part of the docs are commented out in the analysis-core.tex file.