All right: I have rewritten bondGenerator.c into tcl. So from now on, no external scripts in C are needed.
Also I encapsulated the object creation inside procedures. That means, that the user just needs to write one command to create a model for a blood cell. Of course, he needs to supply the mesh files.

The tcl script is calles object_in_fluid.tcl and I put it in scripts. I changed scripts/Makefile.am to include this script in the distribution.

Further I changed the UG accordingly.

Finally, I added a test case. It is located in testsuite/object_in_fluid.tcl. It operates also with the following two files:
object_in_fluid_system.data.init
object_in_fluid_system.data.final.

The description of the testcase is in the tcl file self.

Ivan