I have added the source tex files for object-in-fluidUG located in samples/object-in-fluid/object-in-fluidUG.
I have corrected the NAN in angle_btw_triangles. The problem was that I computed sqrt from a negative value.

Also, I unified the computation of the angle between triangles with tcl script. in scripts/object-in-fluid.tcl, in procedure angle_btw_triangles, there was still old way of computation of the angle. So I changed it accroding to the function in utils.h.

Ivan