Little addittion to wall constraint
From the text, it was not clear that, that the distance is measured in the direction of the normal vector. Although this seems to be clear, if the normal vector points away from the origin, it makes a huge different in the contrary case. Imagine a wall with normal (0,0,-1). In order to place it at z=5, the distance has to be set to -5. 
It is worth emphasizing this.