finalized structure for blockfile2pickle.py
This tools converts tcl blockfiles into appropriate member classes of
the System class. Right now the script supports particles and the
lennard-jones interaction.

Since the tcl interface is as awful as it is, different commands have to
be handled in a case-by-case basis. The script provides enough
infrastructure for that endeavour, so that parsing of different commands
can be implemented if need be.

Specifically for parsing interactions, enough groundwork has been laid
out so that all it takes is to created an appropriate dict for the
corresponding <interaction>.set_params() method.