Provisionally fix dangling boost_comm
This provisionally solves the problem which Tilak described in the developers
meeting: (Arbitrary) segfaults when using multiple processes. This is due to
UB.  Boost::mpi::communicator boost_comm is instantiated in communication.cpp
and copied to MpiCallbacks. However, the underlying comm_cart is later changed
in grid_changed_n_nodes (grid.hpp) which invalidates the MPI_Comm the boost
communicator points to.

This commit is a quick and dirty fix in order to get ESPResSo running, again.
It changes the MpiCallbacks class and the RuntimeErrorCollector to hold a
pointer to a boost::mpi::communicator in order to be able to change the
underlying boost_comm in grid.cpp.  With the current state, one should
absolutely NEVER copy boost_comm, because it can change. However, this problem
should definitely be fixed differently.