find hdf5: testing with this setup
Use a slightly modified `FindHDF5.cmake` module from version 3.3 if the
current cmake version is smaller than 3.4 . System is ubuntu 16.04, with both
HDF5 serial and parallel installed. The customized module is only
included when needed, to avoid renaming it. Cmake doesn't like it if the
find module has a different name than the package that it's looking for.

Test as follows: running cmake and make should build successfully. When
the code which includes the custom module is not present, but the
`HDF5_PREFER_PARALLEL` flag is set, build should fail for cmake < 3.4
and succeed for versions above.