Fix bug of implicit order assumption
The different types of GhostCommunications (GHOST_SEND and GHOST_RECV)
might be ordered differently in the various GhostCommunicators, or they
might be changed by a repartitioning.

Since the commbufs are static, this can lead to sendbufs being used as
recvbufs and, therefore, the bondbufs of receive operations not being
empty although they should and the code depends on them being empty.