ibm: tests for ineritaless tracer vs and tirel, tribend interactions
Based on the tests receaved by email from Stephan Gekle.

The current state has the following issues:
On the cpu, in the triel test, particles seem to swap places in the triangle (short vs long distance)
Not so on the gpu.
On the gpu, the three tests run individually, but not in combination. Cuda memory error lb_inertialess_tracers_cuda.cu:93 (warning about allocation of size 0).