Remove argument of Doxygen \file commands
Doxygen command \file takes a file path as an optional argument. By
default, the input file path is used. It is good practice to leave
the optional argument blank to avoid conflicts. For example, command
`\file initialize.cpp` in file `/src/core/initialize.cpp` will trigger
a warning because 14 files in the project are named `initialize.cpp`,
while `\file src/core/initialize.cpp` won't trigger a warning, but it
must be updated everytime the file is moved or renamed. Providing a
file path is only necessary in specific cases, for example when C++
files are generated with a script via `FILTER_PATTERNS`.