Rename local variables, document bond angles
Renamed variables force1, force2, force3 as f_mid, f_left resp. f_right
to make it clear on which particles the forces are acting, and to be
consistent with the p_mid (particle pointer) and r_mid (particle
position) notation used in these functions.

The angle potentials and the derivation of their forces are now fully
documented in a dedicated page. Other bonded interactions will be
described there as well in the future.