Workaround until CTest implements status "Skipped"
The CMake "PROPERTIES FIXTURES_REQUIRED" fixture only checks if the
dependency passed the test. When the dependency skipped a test, CTest
marks it as "Passed", because it does not implement statuses "Skipped"
nor "Partially Skipped". So a dependency that "Passed" the test may
not have created output files necessary for the dependent test, which
will fail instead of being skipped.