Mitigate I/O race condition in tutorial 07 test
Unclear error on config "empty" causes the second import to fail upon
reading the processed python file that was just written to in the
importlib wrapper. Ordering the import order causes the test to be
skipped on the first import instead of the second import. This
doesn't change the test logic and seems to fix the issue.