Enable grouping of arguments in numpydoc
The napoleon plugin in Sphinx does not allow grouping multiple
arguments on the same line. Grouped arguments didn't get rendered
in the HTML output. The new option fixes that.

The generated espressomd.html only changed in 286 lines out of 10852
lines. Most of the modification are simplifications of the HTML code
and replacement of quotation marks by their typographic equivalent.