Patch Doxygen
When Doxygen parses the arguments of a class method, if an argument
has the same name as another method in the same class, the argument
name is replaced with an empty string, and any \param command
documenting that argument raises a warning.

The CI Doxygen script now silences such warnings if the argument is
found in the generated XML argsstring node of the class method.