Fix SIP bug
To build a new Python Constraint object from a core Constraint
object, a Cython PObjectId object is passed as a named argument
"oid" to the Python Constraint __init__() method. All Constraint
classes must provide an __init__(**kwargs) method without any
positional argument for this to work. This fixes the issues with
checkpointing and visualization of constraints not based on shapes.