Fix for unpatched OpenMPI versions
A bug in OpenMPI versions 2.0-2.1, 3.0.0-3.0.2 and 3.1.0-3.1.2
causes a segmentation fault when running the OpenGL visualizer
with ADDITIONAL_CHECKS enabled together with either DIPOLES or
ELECTROSTATICS. The bug doesn't manifest itself when the enum
values are broadcasted with `static_cast<int>(coulomb.method)`
and `static_cast<int>(dipole.method)`. The OpenMPI bugfix was
not backported on Ubuntu 18.04.