Fix script logic
The original pattern wouldn't work if a filename was the suffix of
another, e.g. `lb.py` would be marked as used in the CMakeLists.txt
because of a match at the line `python_test(FILE engine_lb.py)`.
The new pattern is also not a regular expression: the dot in the
file extension doesn't match any character, only the dot character.