Limit default value of TEST_NP to at most 4 cores
When no default value is provided for TEST_NP, value `nproc / 2 + 1`
is used instead, which can be large on hyperthreaded CPUs. This leads
to issues in parallel tests that don't scale well with a large number
of cores (e.g. when there aren't enough particles) or with a large
odd number of cores (e.g. LB or P3M).