Update array_properties.py
In the ArrayPropertyTest class, the lbf actor setup was moved to the test_lb() method, which was the only test using the actor. The actor is now removed after the test. This fixes an issue in test_partial_periodic() when changing the system's periodicity. The error which occured can be seen in event.cpp, function on_cell_structure_change(). This, however, didn't cause the test to fail because the error was not handled by Python. It did cause the test to fail during development of the Stokesian Dynamics functionality, when a handle_errors() was added into globals.pyx.