Cleanup espressomd.utils imports
Only import what is used. Make a consistent use of `cimport` for
functions/classes declared in .pxd files (even if the body is
defined in the .pyx file) and `import` for functions/classes
declared in .pyx files. `cimport` statements in .pxd files must
also appear in the corresponding .pyx file.