Fix bond connectivity issues at the box boundaries
Calculate how to cut bonds that cross one side of the simulation
box using an analytical solution instead of a heuristic algorithm.
The heuristic was incorrectly configured, leading to wrong bond
angles for cut bonds. This was particularly visible when drawing
periodic images. Also, the old algorithm assumed the box was cubic,
but now all directions are checked.