Remove FindCUDA feature deprecated in CMake 3.10
The feature is not longer required, and actually has compatibility
issues with CUDA 10.1 (e.g. `CUDA_cublas_device_LIBRARY-NOTFOUND`).
From the CMake 3.10 documentation:
["The FindCUDA module has been superseded by first-class support
for the CUDA language in CMake. It is no longer necessary to use
this module or call `find_package(CUDA)`. This module now exists
only for compatibility with projects that have not been ported."](
https://cmake.org/cmake/help/v3.10/module/FindCUDA.html)