Refactor build type and CUDA compiler selection
Generate a warning for incorrect build types and override them with
'Release'. Move the `set(CMAKE_BUILD_TYPE CACHE)` declaration out
of the conditional such that its help message always gets displayed
in cmake/ccmake. List the possible values as properties to allow
cycling in ccmake and cmake-gui (instead of manually typing them).
Same thing for WITH_CUDA_COMPILER. This is achieved by creating a
wrapper around the `option()` function that accepts an enum value
(stored as a string) instead of a boolean value.