CMake: Remove re-definition of CMAKE_CXX_FLAGS
The CMAKE_CXX_FLAGS CMake variable cannot be added directly to an
INTERFACE target. This will duplicate it at the end of the list as
a whole string, which can lead to a compiler error. For example,
`CXXFLAGS="-O3 -Wall" cmake ..` will generate a compilation command
in the form `c++ <flags> -O3 -Wall <flags> "-O3 -Wall" <filenames>`,
which is incorrect (there is no optimization level "3 -Wall"). This
will fail randomly depending on the compiler. This is incompatible
with automated build systems such as RPM Build and OSC.