core: Refactor exception handling in (D)P3M tuning
The old exception mechanism simply printed the error message to a
log string, without halting the flow of the program. This either
left the system in an undefined state, or caused other errors to
be thrown at the end of the tuning functions, making it difficult
to trace the true origin of the issue.