python: Fix error message in LB nodegrid
The np.array() casts the entire array to string if the last element
is a string, causing the assertion in LBFluidRoutines() to complain
about the first element being a string, which it is not in the user
input. The logic for the assertion in HydrodynamicInteraction() can
also be replaced by utils.check_type_or_throw_except().