tests: Fix broken ScaFaCoS test
system.actors.clear() doesn't clear the DLC actor on the first pass.
In that situation, if the DLC check fails, it exits early before the
calls to `del system.actors[0]` and triggers the tearDown() method,
which cannot remove DLC. The system is then in an undetermined state
for the ScaFaCoS check, which throws an exception.