core: Fix broken Utils::Mpi::gather_buffer()
The function must move the first n_elem of the buffer on rank root
to the correct place, otherwise the first n_elem of the buffer on
rank 0 will overwrite them, and a hole of size n_elem can be found
somewhere in the final buffer. This bug doesn't affect code where
root is 0.