CMake: Remove option_if_available() function
The feature is no longer necessary and leads to convoluted
conditional branches. It also causes confusing behavior: the
state of the options is persistent (re-running CMake without
first clearing the build directory leads to a different behavior
compared to running CMake in a fresh build directory), it's not
always clear what will be installed, and the CMake log contains
irrelevant messages about libraries not found for features that
were not selected by the user.

This reverts commit 4f280f66d02d47418d3e256be3e01527d8af2e76.