core: Convert debug statements into assertions
These checks are debug assertions intended to catch obvious errors
during development. They aren't meant to be visible to the user:
the MPI rank number assertion is already checked for when setting
up magnetostatics actors and the flag assertion cannot be triggered
by the user since these flag values are constant expressions in the
calling functions.