core: Fix cell resort bug
When calling CellStructure::resort_particles from the python
callback mpi_resort_particles(), ghost particles are invalidated
and default-constructed, but not updated with the info of the
real particles. This results in undefined behavior later on in
the integration loop.

In the normal flow of the integration loop this is not an issue,
because the method is called from cells_update_ghosts(), which
immediately updates the ghost info.