walberla: Re-use destructor free function
The placement operator new() is safe against constructors throwing
exceptions. When this happens, operator new() frees the allocated
memory via operator delete() and re-throws the exception. The
assignment operator isn't executed and the variable contains the
old value (in our case nullptr), making it safe to call operator
delete() on the variable.
https://en.cppreference.com/w/cpp/language/new#Construction