python: Forward C++ exceptions to Python
Intermediate Cython functions need to be marked as `except +` to
forward a C++ exception from the core. If the intermediate Cython
function has a return type, then the returned value is not a Python
type and the exception cannot be wrapped in the returned object,
instead a sentinel value is used for C++ types that support it,
while void functions need to be marked as `except *` since they
don't return a value, but this comes with an overhead in the form
of a function call to `PyErr_Occurred()`.

For simplicity, the return types of all intermediate Cython
functions were removed in the LB interface.