samples: Fix broken LB Poiseuille flow script
The observable was sampling every 100 points, but the LB fluid
was only 8 grid points in diameter. This caused a substantial
overhead, especially with GPU acceleration due to the frequent
interpolation (runtime: 2 min on coyote11, 6 min on coyote 1).
The sampling was also too short to reach convergence, resulting
in a plug flow profile instead of a Poiseuille flow profile.