walberla: Refactor LB fluid and lattice infrastructure
Use the opaque pointer idiom to hide the walberla blockforest
header files from the EspressoCore and ScriptInterface. Make
the LB class a pseudo-actor class that is also a script interface
object with no parameters; construction is deferred until the
call to self.call_method('init_lb_walberla') when the LB "actor"
is activated. Create a new script interface class LatticeWalberla
to manage a shared pointer to the core class that contains a
shared pointer to the blockforest.