walberla: Guard against reference counting bug
Properly deallocate script interface objects when the list of actors
is destroyed by the python garbage collector. This prevents undefined
behavior from a known bug in waLBerla's custom reference counting
mechanism that miscalculates the number of references to an object.
Without this patch, in any LB simulation where VTK objects are added
to the VTK containers of the LBWalberlaImpl class, at system exit,
the walberla::vtk::VTKOutput::~VTKOutput() destructor is called twice.