python: Clear the list of actors upon exit
Actor objects can set up global bitfields as a side-effect of their
construction. Sanity checks in the core depend on these bitfields,
so it is necessary to call the Actor destructors from python when
the system exits, so as to reset global bitfields in reverse order.
This avoids accidental side effects, such as throwing an exception
in the destructor of another Actor object that doesn't allow changes
to the MD cellsystem.