core: Reduce size of serialized Array and Vector
Since the size of Array and Vector objects is known at compile time,
it doesn't have to be communicated. This spares an integer and a
prefix code during MPI communication operations, thus reducing the
serialization buffer size by 33% for Vector3i (from 24 to 16 bytes)
and by 22% for Vector3d (from 36 to 28 bytes).