CMake: Don't rename test files when possible
When a test case doesn't need CMake variable substitution during
configuration, there is no need to rename it. Renaming test files
hinders Coverage.py's ability to map test files in the build
directory to the corresponding test files in the root directory.
Use FILE_SUFFIX to change the filepath during configuration, and
SUFFIX to change the test name in CTest.