CMake: Put binary targets in namespace `Espresso::`
This prevents target collision when including external libraries.
To take the example of shapes, which use the Pitchfork layout:
Espresso_shape is the binary target which can only be modified
by src/shapes/CMakeLists.txt, while the corresponding alias
Espresso::shape is a const reference to the binary target,
which has to be used by consumers of the target.