doc: Improve Widom tutorial
Provide the code for the matplotlib figure and the main simulation
loop. Detail the derivation of the analytical formula for the excess
potential (DH and Davies) and source the relevant chapter in Atkins'
textbook. Reduce the simulation time in the testsuite and make the
test deterministic by removing all but one degree of freedom in the
P3M algorithm. Flush stdout to avoid breaking stderr (tqdm progress
bars). Fix issues in LaTeX equations.