Propagate exceptions during type conversion (#4911)
Release 4.2.0 generalized the use of utility functions to convert
Python vector-valued objects of size 3 to the corresponding C++
script interface Vector3d or Vector3i types. Exceptions raised in
these functions were not propagated to the Python interpreter,
printing "Exception ignored in: 'espressomd.utils.make_Vector3d'"
to the screen and leaving the call site with a fully or partially
uninitialized object. This is now resolved.